Oppdater modal for endringstidspunkt og legg på feilmelding
