+++ src/frontend/komponenter/Fagsak/Vedtak/VedtakBegrunnelserTabell/endringstidspunkt/useOppdaterEndringstidspunktSkjema.ts
@@ -1,13 +1,20 @@
 import { useEffect } from 'react';
 
 import type { ISODateString } from '@navikt/familie-datovelger';
-import { feil, useFelt, useSkjema, ok } from '@navikt/familie-skjema';
 import type { FeltState } from '@navikt/familie-skjema';
+import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 
-import type { IBehandling } from '../../../../typer/behandling';
-import type { FamilieIsoDate } from '../../../../utils/kalender';
-import { erFør, erIsoStringGyldig, kalenderDato } from '../../../../utils/kalender';
-import { iDag, KalenderEnhet, leggTil, serializeIso8601String } from '../../../../utils/kalender';
+import type { IBehandling } from '../../../../../typer/behandling';
+import type { FamilieIsoDate } from '../../../../../utils/kalender';
+import {
+    erFør,
+    erIsoStringGyldig,
+    iDag,
+    kalenderDato,
+    KalenderEnhet,
+    leggTil,
+    serializeIso8601String,
+} from '../../../../../utils/kalender';
 
 const STANDARD_ANTALL_DAGER_FRIST = 3 * 7;
 
@@ -27,7 +34,7 @@ export const useOppdaterEndringstidspunktSkjema = (
     >({
         felter: {
             endringstidspunkt: useFelt<FamilieIsoDate | undefined>({
-                verdi: endringstidspunkt ?? standardfrist,
+                verdi: endringstidspunkt,
                 valideringsfunksjon: validerEndringstidspunkt,
             }),
         },
